api/requirements.txt
appdirs==1.4.3
asgiref==3.2.10
autopep8==1.5.3
CacheControl==0.12.6
certifi==2019.11.28
chardet==3.0.4
colorama==0.4.3
contextlib2==0.6.0
distlib==0.3.0
distro==1.4.0
dj-database-url==0.5.0
Django==3.0.5
django-filter==2.3.0
django-mongodb-engine==0.6.0
djangorestframework==3.11.0
gunicorn==20.0.4
html5lib==1.0.1
idna==2.8
ipaddr==2.2.0
lockfile==0.12.2
Markdown==3.2.2
msgpack==0.6.2
netifaces==0.10.9
packaging==20.3
pep517==0.8.2
pep8==1.7.1
prettytable==0.7.2
progress==1.5
psycopg2-binary==2.8.5
pycodestyle==2.6.0
PyJWT==1.7.1
pyparsing==2.4.6
python-dotenv==0.14.0
pytoml==0.1.21
pytz==2020.1
requests==2.22.0
retrying==1.3.3
six==1.14.0
sqlparse==0.2.4
toml==0.10.1
twilio==6.44.0
urllib3==1.25.8
webencodings==0.5.1
whitenoise==5.1.0
